re-refactor into functional modules
