+++ bct/bct/__init__.py
@@ -1,13 +1,2 @@
-from .centrality import *
-from .clustering import *
-from .modularity import *
-from .core import *
-from .degree import *
-from .distance import *
-from .motifs import *
-from .other import *
-from .physical_connectivity import *
-from .reference import *
-from .similarity import *
-from .visualization import *
-from .miscellaneous_utilities import *
+from .algorithms import *
+from .utils import *
