+++ bct/bct/algorithms/__init__.py
@@ -1,0 +1,10 @@
+from .centrality import *
+from .clustering import *
+from .modularity import *
+from .core import *
+from .degree import *
+from .distance import *
+from .motifs import *
+from .physical_connectivity import *
+from .reference import *
+from .similarity import *
