+++ bct/bct/algorithms/core.py
