+++ bct/bct/algorithms/motifs.py
@@ -1,7 +1,7 @@
 from __future__ import division
 import numpy as np
-from .miscellaneous_utilities import BCTParamError
-from .other import binarize
+from bct.bct.utils.miscellaneous_utilities import BCTParamError
+from bct.bct.utils.other import binarize
 
 motiflib='motif34lib.mat'
