+++ bct/bct/algorithms/reference.py
@@ -1,8 +1,8 @@
 from __future__ import division
 import numpy as np
-from .miscellaneous_utilities import BCTParamError
+from bct.bct.utils.miscellaneous_utilities import BCTParamError
 from .clustering import number_of_components
-from .other import binarize
+from bct.bct.utils.other import binarize
 
 
 def latmio_dir_connected(R,iter,D=None):
