+++ bct/bct/utils/__init__.py
@@ -1,0 +1,3 @@
+from .miscellaneous_utilities import *
+from .visualization import *
+from .other import *
